extract FunctionRouterExecutorFactory class
