+++ activiti-cloud-service-common/activiti-cloud-service-messaging-config/src/main/java/org/activiti/cloud/common/messaging/config/FunctionRouterConfiguration.java
@@ -17,19 +17,13 @@ package org.activiti.cloud.common.messaging.config;
 
 import static org.activiti.cloud.common.messaging.config.CompletableFutureRetry.supplyAsyncWithRetry;
 
-import jakarta.annotation.PreDestroy;
-import java.time.Duration;
 import java.util.Collection;
 import java.util.List;
 import java.util.Map;
 import java.util.Optional;
 import java.util.concurrent.CompletableFuture;
 import java.util.concurrent.CompletionException;
-import java.util.concurrent.ConcurrentHashMap;
-import java.util.concurrent.ExecutionException;
 import java.util.concurrent.ExecutorService;
-import java.util.concurrent.Executors;
-import java.util.concurrent.TimeUnit;
 import java.util.function.BiConsumer;
 import java.util.function.BiFunction;
 import java.util.function.Consumer;
@@ -142,8 +136,8 @@ public class FunctionRouterConfiguration {
 
     @Bean
     @ConditionalOnMissingBean
-    Function<String, ExecutorService> functionExecutorFactory() {
-        return new FunctionExecutorFactory();
+    Function<String, ExecutorService> functionRouterExecutorFactory() {
+        return new FunctionRouterExecutorFactory();
     }
 
     @Bean
@@ -166,12 +160,12 @@ public class FunctionRouterConfiguration {
     @Bean
     Function<Message<?>, ExecutorService> functionExecutorSelector(
         Function<Message<?>, String> functionRegistrationSelector,
-        Function<String, ExecutorService> functionExecutorFactory
+        Function<String, ExecutorService> functionRouterExecutorFactory
     ) {
         return new Function<>() {
             @Override
             public ExecutorService apply(Message<?> message) {
-                return functionRegistrationSelector.andThen(functionExecutorFactory).apply(message);
+                return functionRegistrationSelector.andThen(functionRouterExecutorFactory).apply(message);
             }
         };
     }
@@ -361,81 +355,5 @@ public class FunctionRouterConfiguration {
                 return bean;
             }
         };
-    }
-
-    public static class FunctionExecutorFactory implements Function<String, ExecutorService> {
-
-        private final Map<String, ExecutorService> executors = new ConcurrentHashMap<>();
-        private Duration timeout = Duration.ofSeconds(5);
-
-        private final Function<String, ExecutorService> executorServiceFactory = registration ->
-            Executors.newSingleThreadScheduledExecutor(runnable -> {
-                final var thread = new Thread(runnable);
-                thread.setName(registration);
-
-                return thread;
-            });
-
-        public FunctionExecutorFactory() {}
-
-        @Override
-        public ExecutorService apply(String key) {
-            return executors.computeIfAbsent(key, executorServiceFactory);
-        }
-
-        @PreDestroy
-        public void destroy() {
-            try {
-                shutdown();
-
-                if (!awaitTermination(timeout.getSeconds(), TimeUnit.SECONDS)) {
-                    shutdownNow();
-                }
-            } catch (InterruptedException e) {
-                Thread.currentThread().interrupt();
-                shutdownNow();
-            } finally {
-                executors.clear();
-            }
-        }
-
-        public void shutdown() {
-            executors.values().forEach(ExecutorService::shutdown);
-        }
-
-        public void shutdownNow() {
-            executors.values().forEach(ExecutorService::shutdownNow);
-        }
-
-        public boolean awaitTermination(final long timeout, final TimeUnit unit) throws InterruptedException {
-            final var cfs = executors
-                .values()
-                .stream()
-                .map(executor ->
-                    CompletableFuture.supplyAsync(() -> {
-                        try {
-                            return executor.awaitTermination(timeout, unit);
-                        } catch (InterruptedException ignored) {
-                            Thread.currentThread().interrupt();
-                        }
-
-                        return false;
-                    })
-                )
-                .toList();
-
-            try {
-                return CompletableFuture
-                    .allOf(cfs.toArray(CompletableFuture[]::new))
-                    .thenApply(v -> cfs.stream().map(CompletableFuture::join).allMatch(Boolean.TRUE::equals))
-                    .get();
-            } catch (ExecutionException e) {
-                throw new InterruptedException();
-            }
-        }
-
-        public void setTimeout(Duration timeout) {
-            this.timeout = timeout;
-        }
     }
 }
